requirements.txt
discord.py==1.7.3
python_pushsafer==1.0
requests==2.26.0
